feat: placed in separate modules os/navigation
